+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/completion/properties/ProjectConfigurationPropertiesLoader.kt
@@ -18,24 +18,13 @@
 package com.explyt.spring.core.completion.properties
 
 import com.explyt.module.ExternalSystemModule
-import com.explyt.spring.core.JavaCoreClasses
-import com.explyt.spring.core.SpringCoreClasses
 import com.explyt.spring.core.SpringProperties.ADDITIONAL_CONFIGURATION_METADATA_FILE_NAME
 import com.explyt.spring.core.SpringProperties.HINTS
 import com.explyt.spring.core.SpringProperties.PROPERTIES
 import com.explyt.spring.core.completion.properties.utils.ProjectConfigurationPropertiesUtil
-import com.explyt.spring.core.service.SpringSearchService
 import com.explyt.spring.core.util.PropertyUtil
-import com.explyt.spring.core.util.SpringCoreUtil.isMimeTypeClass
-import com.explyt.util.ExplytPsiUtil.isFinal
-import com.explyt.util.ExplytPsiUtil.isMap
-import com.explyt.util.ExplytPsiUtil.isMetaAnnotatedBy
-import com.explyt.util.ExplytPsiUtil.isNonStatic
-import com.explyt.util.ExplytPsiUtil.resolvedPsiClass
 import com.explyt.util.ExplytPsiUtil.returnPsiClass
-import com.explyt.util.ExplytPsiUtil.returnPsiType
 import com.explyt.util.runReadNonBlocking
-import com.intellij.codeInsight.AnnotationUtil
 import com.intellij.json.psi.JsonFile
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.project.Project
@@ -91,7 +80,7 @@ class ProjectConfigurationPropertiesLoader(project: Project) : AbstractSpringMet
                 else -> null
             } ?: continue
 
-            collectConfigurationProperty(
+            PropertyUtil.collectConfigurationProperty(
                 module,
                 configurationPropertiesType,
                 configurationPropertiesType,
@@ -111,167 +100,12 @@ class ProjectConfigurationPropertiesLoader(project: Project) : AbstractSpringMet
         return result
     }
 
-    private fun collectConfigurationProperty(
-        module: Module,
-        ownerConfigurationClass: PsiClass,
-        targetClass: PsiClass,
-        prefix: String,
-        result: MutableMap<String, ConfigurationProperty>
-    ) {
-        val nestedFields = getNestedPropertyWrappers(targetClass)
-        val finalFields = getFieldPropertyWrappers(targetClass)
-        val setterMethods = getMethodPropertyWrappers(module, targetClass, nestedFields)
-
-        for (it in nestedFields) {
-            val propertyTypeClass = it.psiType.resolvedPsiClass
-            if (propertyTypeClass != null) {
-                collectConfigurationProperty(
-                    module,
-                    ownerConfigurationClass,
-                    propertyTypeClass,
-                    "$prefix.${it.name}",
-                    result
-                )
-            }
-        }
-
-        val propertyWrappers = finalFields + setterMethods
-        for (propertyWrapper in propertyWrappers) {
-            val propertyName = propertyWrapper.name
-            val psiType = propertyWrapper.psiType
-
-            if (psiType is PsiClassType) {
-                val propertyTypeClass = psiType.resolve()
-                val javaFile = propertyTypeClass?.containingFile as? PsiJavaFile ?: continue
-
-                if (javaFile.packageName != JavaCoreClasses.PACKAGE_JAVA_LANG &&
-                    javaFile.packageName != JavaCoreClasses.PACKAGE_KOTLIN
-                ) {
-                    collectConfigurationProperty(
-                        module,
-                        ownerConfigurationClass,
-                        propertyTypeClass,
-                        "$prefix.$propertyName",
-                        result
-                    )
-                }
-            }
-
-            val name = "$prefix.$propertyName"
-            result[name] = ConfigurationProperty(
-                name,
-                ConfigurationPropertiesLoader.getPropertyType(propertyWrapper.psiType),
-                propertyWrapper.type,
-                propertyWrapper.sourceType,
-                propertyWrapper.description,
-                propertyWrapper.default,
-                propertyWrapper.deprecation
-            )
-        }
-    }
-
-    private fun getMethodPropertyWrappers(
-        module: Module,
-        targetClass: PsiClass,
-        nestedFields: List<FieldPropertyWrapper>
-    ): List<MethodPropertyWrapper> {
-        val result = mutableListOf<MethodPropertyWrapper>()
-        val setterMethods = PropertyUtil.getSetterMethods(targetClass, nestedFields)
-        val getterMethods = PropertyUtil.getGetterMethods(targetClass, nestedFields)
-        for (method in setterMethods) {
-            val setterIdentifier = getIdentifierFromSetterMethod(method) ?: continue
-            var methodPropertyWrapper: MethodPropertyWrapper? = null
-            for (it in getterMethods) {
-                val getterIdentifier = getIdentifierFromGetterMethod(it)
-                if (getterIdentifier!=null && getterIdentifier.text == setterIdentifier.text) {
-                    val deprecation = getDeprecationInfo(module, it)
-                    methodPropertyWrapper = MethodPropertyWrapper(method, deprecation)
-                }
-            }
-            if (methodPropertyWrapper == null) {
-                methodPropertyWrapper = MethodPropertyWrapper(method, null)
-            }
-            result += methodPropertyWrapper
-        }
-        return result
-    }
-
-    private fun getNestedPropertyWrappers(targetClass: PsiClass): List<FieldPropertyWrapper> {
-        return targetClass.allFields
-            .filter { it.isMetaAnnotatedBy(SpringCoreClasses.NESTED_CONFIGURATION_PROPERTIES) }
-            .map { FieldPropertyWrapper(it) }
-    }
-
-    private fun getFieldPropertyWrappers(targetClass: PsiClass): List<FieldPropertyWrapper> {
-        val fields = if (targetClass.isEnum) {
-            emptyList()
-        } else if (targetClass.isMimeTypeClass()) {
-            targetClass.allFields.filter { it.returnPsiType?.isMap == true }
-        } else {
-            targetClass.allFields.filter {
-                it.isNonStatic
-                        && it.isFinal
-                        && !it.isMetaAnnotatedBy(SpringCoreClasses.NESTED_CONFIGURATION_PROPERTIES)
-            }
-        }
-        return fields.map { FieldPropertyWrapper(it) }
-    }
 
     private fun findMetadataFiles(module: Module): List<PsiFile> {
         val sourceMeta = ExternalSystemModule.of(module).sourceMetaInfDirectory ?: return emptyList()
         return sourceMeta.files.asSequence()
             .filterNotNull()
             .filter { ADDITIONAL_CONFIGURATION_METADATA_FILE_NAME.equals(it.name, true) }.toList()
-    }
-
-    private fun getIdentifierFromSetterMethod(psiMethod: PsiMethod): PsiIdentifier? {
-        val codeBlock = psiMethod.childrenOfType<PsiCodeBlock>()
-        if (codeBlock.isEmpty()) {
-            return psiMethod.identifyingElement as? PsiIdentifier
-        }
-        val expressions = codeBlock[0].childrenOfType<PsiExpressionStatement>()
-        val assignmentExpressions = expressions.asSequence()
-            .map { it.childrenOfType<PsiAssignmentExpression>() }
-            .filter { it.isNotEmpty() }
-            .toList()
-        return assignmentExpressions
-            .asSequence()
-            .flatMap { it.asSequence() }
-            .filter { it.lastChild != null }
-            .filter {
-                it.lastChild.text == psiMethod.parameterList.parameters[0].lastChild?.text
-                        || it.lastChild.text == psiMethod.parameterList.parameters[0].name
-            }
-            .map { it.firstChild }
-            .filterIsInstance<PsiReferenceExpression>()
-            .map { it.childrenOfType<PsiIdentifier>().firstOrNull() }
-            .firstOrNull { it != null }
-    }
-
-    private fun getIdentifierFromGetterMethod(psiMethod: PsiMethod): PsiIdentifier? {
-        val codeBlock = psiMethod.childrenOfType<PsiCodeBlock>()
-        if (codeBlock.isEmpty()) {
-            return psiMethod.identifyingElement as? PsiIdentifier
-        }
-        val expressions = codeBlock.flatMap { it.childrenOfType<PsiReturnStatement>() }
-
-        return expressions.lastOrNull()
-            ?.childrenOfType<PsiReferenceExpression>()?.lastOrNull()
-            ?.childrenOfType<PsiIdentifier>()?.lastOrNull()
-    }
-
-    private fun getDeprecationInfo(module: Module, method: PsiMethod?): DeprecationInfo? {
-        if (method == null || !method.isDeprecated) {
-            return null
-        }
-        val metaHolder = SpringSearchService.getInstance(module.project)
-            .getMetaAnnotations(module, SpringCoreClasses.DEPRECATED_CONFIGURATION_PROPERTIES)
-        val reasonAnnotation = metaHolder.getAnnotationMemberValues(method, setOf("reason")).firstOrNull()
-        val replacementAnnotation = metaHolder.getAnnotationMemberValues(method, setOf("replacement")).firstOrNull()
-        val reason = if (reasonAnnotation != null) AnnotationUtil.getStringAttributeValue(reasonAnnotation) else null
-        val replacement =
-            if (replacementAnnotation != null) AnnotationUtil.getStringAttributeValue(replacementAnnotation) else null
-        return DeprecationInfo(DeprecationInfoLevel.WARNING, replacement, reason)
     }
 }
 
@@ -343,7 +177,8 @@ class FieldPropertyWrapper(psiField: PsiField, deprecationInfo: DeprecationInfo?
     override val deprecation: DeprecationInfo? = deprecationInfo
 }
 
-private class MethodPropertyWrapper(psiMethod: PsiMethod, deprecationInfo: DeprecationInfo? = null) : PropertyWrapper<PsiMethod>(psiMethod) {
+class MethodPropertyWrapper(psiMethod: PsiMethod, deprecationInfo: DeprecationInfo? = null) :
+    PropertyWrapper<PsiMethod>(psiMethod) {
 
     override val psiType: PsiType = psiMember.parameterList.parameters[0].type