+++ modules/spring-core/src/test/kotlin/com/explyt/spring/core/reference/kotlin/YamlReferenceTest.kt
@@ -21,7 +21,9 @@ import com.explyt.spring.core.properties.providers.ConfigKeyPsiElement
 import com.explyt.spring.core.properties.providers.ConfigurationPropertyKeyReference
 import com.explyt.spring.test.ExplytKotlinLightTestCase
 import com.explyt.spring.test.TestLibrary
+import com.intellij.lang.properties.psi.impl.PropertiesFileImpl
 import com.intellij.psi.impl.source.resolve.reference.impl.PsiMultiReference
+import com.intellij.psi.impl.source.resolve.reference.impl.providers.FileReference
 
 class YamlReferenceTest : ExplytKotlinLightTestCase() {
     override fun getTestDataPath(): String = super.getTestDataPath() + "reference/properties"
@@ -33,7 +35,7 @@ class YamlReferenceTest : ExplytKotlinLightTestCase() {
             TestLibrary.resilience4j_2_2_0,
         )
 
-    fun testRefInnerClass() {
+    fun testRefKeyInnerClass() {
         myFixture.copyFileToProject("LssConfigurationProperties.kt")
         myFixture.configureByText(
             "application.yaml",
@@ -56,7 +58,7 @@ lss:
         assertEquals(nameClass, "setLssPlanConfiguration")
     }
 
-    fun testRefInnerClassField() {
+    fun testRefKeyInnerClassField() {
         myFixture.copyFileToProject("LssConfigurationProperties.kt")
         myFixture.configureByText(
             "application.yaml",
@@ -80,8 +82,7 @@ lss:
         assertEquals(name, "setExact")
     }
 
-    // in the next commit
-    fun _testRefInnerClassFieldBoolean() {
+    fun testRefKeyInnerClassFieldBoolean() {
         myFixture.copyFileToProject("LssConfigurationProperties.kt")
         myFixture.configureByText(
             "application.yaml",
@@ -104,4 +105,33 @@ lss:
         val name = (resolveResult.element as? ConfigKeyPsiElement)?.name
         assertEquals(name, "setExact")
     }
+
+    fun testRefValueResource() {
+        myFixture.copyFileToProject("MainFooProperties.kt")
+        myFixture.configureByText(
+            "application-default.properties",
+            "main.foo-bean-component=fooBeanComponent"
+        )
+        myFixture.configureByText(
+            "application.yaml",
+            """
+main:
+  local:
+    code-resource: classpath:application-de<caret>fault.properties                
+            """.trimIndent()
+        )
+
+        val ref = (file.findReferenceAt(myFixture.caretOffset) as? PsiMultiReference)
+            ?.references?.asSequence()
+            ?.mapNotNull {
+                it as? FileReference
+            }?.firstOrNull()
+
+        assertNotNull(ref)
+        val multiResolve = ref!!.multiResolve(true)
+        assertEquals(1, multiResolve.size)
+        val name = (multiResolve[0].element as? PropertiesFileImpl)?.name
+        assertEquals(name, "application-default.properties")
+    }
+
 }